Plot EOW in Waterbodies - Tidy up
* Not happy tho as had to copy a method (toggleFilterButton()).  Its
innocuous and the lesser of two evils.
